[feat] #85: 홈 추가 화면 데이터 sealed class 추가
