--- app/src/main/java/com/example/findu/domain/model/extra/VolunteerWork.kt
@@ -1,10 +1,0 @@
-package com.example.findu.domain.model.extra
-
-data class VolunteerWork(
-    val institution: String,
-    val recruitmentPeriod: String,
-    val address: String,
-    val workPeriod: String,
-    val workTime: String,
-    val webLink: String
-)
